+++ packages/playwright-core/src/server/firefox/protocol.d.ts
@@ -8,6 +8,10 @@ export module Protocol {
       browserContextId?: string;
       openerId?: string;
     };
+    export type UserPreference = {
+      name: string;
+      value: any;
+    };
     export type CookieOptions = {
       name: string;
       value: string;
@@ -71,6 +75,10 @@ export module Protocol {
     }
     export type enableParameters = {
       attachToDefaultContext: boolean;
+      userPrefs?: {
+        name: string;
+        value: any;
+      }[];
     };
     export type enableReturnValue = void;
     export type createBrowserContextParameters = {
